Added a couple of hacks for CORS/mixed content
Configured a reverse proxy at www.ag7if.net/api that would be a secure
endpoint for the app to hit on the way out to the USNO API (since USNO
does not use SSL). Well, secure until it hits the server and gets passed
on to USNO. This is extremely hacky, but it works for now.

To resolve the CORS problems with the ADDS request, yet even another
proxy was set up through Nginx in order to add the
Access-Control-Allow-Origin header before passing the response back to
the app. I feel less bad about this workaround, but it's still pretty
hacky.