Added a hack to get around mixed content problems
Configured a reverse proxy at api.the-hawk.us that would be a secure
endpoint for the app to hit on the way out to the USNO API (since USNO
does not use SSL). The app's url has been updated to use the reverse
proxy instead of directly using the USNO endpoint. This is extremely
hacky, but it works.